% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download Redcap Report}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{Character string specifying the name of the API token in the user's .Renviron file.}

\item{redcapUrl}{Character string specifying the URL of the Redcap server.}

\item{redcapReportId}{Numeric specifying the ID of the Redcap report to retrieve.}
}
\value{
A tibble containing the contents of the specified Redcap report.
}
\description{
Retrieve a specified Redcap report from a given URL using an API token.
}
\details{
This function uses \code{Sys.getenv()} to read an API token provided as a string (\code{redcapTokenName}) from the user's .Renviron file. It then retrieves the specified Redcap Report (\code{redcapReportId}) from the given URL (\code{redcapUrl}) and returns the contents as a tibble.
}
\examples{
\dontrun{
# Download Redcap report with specified parameters
redcapTokenName = "redcapTokenName"
redcapServerUrl = "https://redcap.emory.edu/api/"
reportId = 46524
myReport = downloadRedcapReport(redcapToken, redcapServerUrl, reportId)
}

}
\references{
Redcap API documentation: \url{https://redcap.emory.edu/api/}
}
\seealso{
\code{\link{httr::POST}}, \code{\link{httr::content}}
}
\concept{Redcap}
\concept{Redcap Functions}
\concept{httr}
\keyword{data}
