% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Perform principal component analysis and return an approximation of the data given a speecified number of PCs}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{Input data as numeric matrix or data frame}

\item{npc}{Integer specifying the number of principal components retained in approximation.}
}
\value{
Numeric matrix approximating the input data based on the specified number of principal components; scaled and centered to patch input data.
}
\description{
This function performs principal component analysis (PCA) on input data \code{x},
then for a specified number of principal components (\code{npc}), returns returns an approximation of the data.
The resulting approximation is rescaled and centered to match the original scaling and centering of the input data.
}
\details{
This function first scales and centers the input data using the \code{scale()} function with \code{center = TRUE} and \code{scale = TRUE}.
Then, it performs PCA using the \code{prcomp()} function and extracts the number of principal components specified by \code{npc}.
The scaled and centered input data is projected onto these principal components to obtain an initial approximation (\code{x_est}).
The function then unscales the approximation (\code{x_est}) using the original scaling and centering parameters (\code{mu} and \code{sigma})
obtained from the input data.
The resulting approximation (\code{x_est_rescaled}) is a rescaled and centered version of the PCA-based approximation of the input data.
When npc is equivalent to the number of columns in \code{x}, then the function output \code{x_est_rescaled} should match \code{x}
}
\examples{
# Generate example data
set.seed(123)
x = matrix(rnorm(100), nrow = 10)  # Example 10x10 matrix

# Perform PCA-based approximation
approx_data = pcApprox(x, npc = 3)

}
