% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize variable names to consistent format: small camel case}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A dataframe or tibble with variable names that may require standardization.}
}
\value{
A dataframe identical to \code{data}, but with standardized variable names in small camel case.
}
\description{
StandardizeNames converts the variable names in the input dataframe or tibble \code{data} to "small camel" case
using the janitor package for cleaning and name standardization.
}
\examples{
data = data.frame("First.Column" = 1:5, "Second_Column" = 6:10)
data_renames = standardizeNames(data)

}
