% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Calculate and plot a Kaplan-Meier survival curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A numerical vector indicating event status: 1 for event occurred, 0 for censored participant.}

\item{time}{A numerical vector representing survival times corresponding to event status. Arbitrary units.}
}
\value{
A \code{ggplot} object displaying the Kaplan-Meier survival curve such that axis labeling, legends can be refined
}
\description{
\code{survCurv} calculates and plots a Kaplan-Meier survival curve (\eqn{S(t)})
based on two independent data arrays of binary event status (1 = event; 0 = censored) and survival time data.
Assumption is that each total rows correspond with total participants, so may need to pad inputs if participants do no experience events or are not censored.
}
\examples{
# Load example data (assuming data is available at the provided URL)
data = read.csv("https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv")
status = data$status
time = data$time

# Calculate and plot survival curve
survCurv(data$status, data$time)

}
