% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale a vector that was originally scaled using 'scale' function}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector that has been scaled and centered using the \code{scale} function.}
}
\value{
The original vector before any scaling and centering.
}
\description{
\code{unscale} takes a given scaled vector \code{x} that has been centered and scaled
(assumed using the \code{scale} function) and reverses the centering and scaling, if any.
}
\details{
This function assumes that the input vector \code{x} was previously scaled
using the \code{scale} function in R. It retrieves the mean (\code{mu}) and standard deviation
(\code{sigma}) used for scaling and centering, then reverses centering and scaling operations to obtain
the original vector.
}
\examples{
set.seed(123)
x = rnorm(10)
scaled_x = scale(x)
original_x = unscale(scaled_x)
all.equal(original_x, x)  # Should return TRUE

}
